feat(Synthwave): move postprocessing to component
