+++ components/content/synthwave/index.vue
@@ -1,10 +1,6 @@
 <script lang="ts" setup>
-import { SRGBColorSpace, WebGLRenderer, Camera, Mesh, Scene, CineonToneMapping, Vector2 } from 'three'
-import { shallowRef } from 'vue'
-import { BloomEffect, FXAAEffect, ScanlineEffect, EffectComposer, RenderPass, EffectPass } from 'postprocessing'
-import { useWindowSize } from '@vueuse/core'
+import { SRGBColorSpace, CineonToneMapping } from 'three'
 import { PALETTE } from './components/palette'
-// Source: https://lospec.com/palette-list/synthwave-9
 
 const gl = {
   clearColor: PALETTE[0],
@@ -13,52 +9,11 @@ const gl = {
   toneMapping: CineonToneMapping,
   logarithmicDepthBuffer: true,
 }
-
-const setupRef = shallowRef(null)
-const disableRenderRef = shallowRef(false)
-let effectComposer: EffectComposer
-
-function setup(renderer: WebGLRenderer, scene: Scene, camera: Camera) {
-  disableRenderRef.value = true
-  effectComposer = new EffectComposer(renderer)
-  const { width, height } = useWindowSize()
-  renderer.setSize(width.value, height.value)
-  renderer.setPixelRatio(Math.min(window.devicePixelRatio, 2))
-  const bloomEffect = new BloomEffect({
-    intensity: 0.5,
-    luminanceThreshold: 0,
-    radius: 4,
-  })
-  const fxaaEffect = new FXAAEffect()
-
-  fxaaEffect.minEdgeThreshold = 0.01
-  const scanline = new ScanlineEffect()
-
-  const renderPass = new RenderPass(scene, camera)
-  const effectPass = new EffectPass(camera, fxaaEffect, bloomEffect, scanline)
-
-  effectComposer = new EffectComposer(renderer)
-  effectComposer.addPass(renderPass)
-  effectComposer.addPass(effectPass)
-}
-
-useRenderLoop().onLoop(({ delta, elapsed }) => {
-  effectComposer?.render()
-})
-
-watch([setupRef], () => {
-  if (setupRef.value) {
-    const mesh = setupRef.value as Mesh
-    mesh.onAfterRender = (renderer, scene, camera, geometry, material, group) => {
-      setup(renderer, scene, camera)
-      mesh.removeFromParent()
-    }
-  }
-})
 </script>
 
 <template>
-  <TresCanvas v-bind="gl" :disable-render="disableRenderRef">
+  <TresCanvas v-bind="gl" :disable-render="true">
+    <Postprocessing />
     <TresWebGLRenderer ref="rendererRef"></TresWebGLRenderer>
     <TresMesh ref="setupRef" :position="[0, 0, 0]"></TresMesh>
